apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
boto3==1.35.36
celery==5.4.0
redis==5.1.1
tenacity==9.0.0
PyJWT==2.9.0
